id double

+++ Program.cs
@@ -114,7 +114,7 @@ app.MapPost("/token",
 //            : Results.NotFound());
 
 app.MapGet("/api/customer/status/{status}", [Authorize] async (MetaContext db, int status) => await db.CombinedTables.Where(a => a.Status == (status == 0 ? false : true)).ToListAsync());
-app.MapGet("/api/customer/{id}", [Authorize] async ( MetaContext db, int id) => await db.CombinedTables.Where(a => a.DataId == id).ToListAsync());
+app.MapGet("/api/customer/{id}", [Authorize] async ( MetaContext db, double id) => await db.CombinedTables.Where(a => a.DataId == id).ToListAsync());
 //app.MapPost("/api/marketing/setStatus", [Authorize] async (MetaContext db, HttpContext context, List<double> idList) =>
 //{
 //    if (context.Request.HasJsonContentType())
